File selection dialog for Import and Demo test recorders
Do not remove the first 5 characters of the filename
as this prevents any file from being opened.
Ungrey the "File info" en "File size" fields while keeping them readonly
by changing the types to GroupSetting and by not disabling the fields.
Change the FormatSize parameter from int to int64_t to fix the file
size display of large files.

+++ mythtv/libs/libmythui/mythuifilebrowser.h
@@ -112,7 +112,7 @@ class MUI_PUBLIC MythUIFileBrowser : public MythScreenType
     void updateWidgets(void);
 
     static bool IsImage(QString extension);
-    static QString FormatSize(int size);
+    static QString FormatSize(int64_t size);
 
     bool               m_isRemote        {false};
